xcity_ai/pyproject.toml
[project]
name = "xcity-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
]
